// src/app/models/Help.ts
export class Help{
    victim: string= ''; 
    reason: Array<string>=[];
    agressor: Array<string>= []; 
    agressorGender: string = '';  
    age: number=0;
    gender: string= ''; 
    place: Array<string> = []; 
    name: string= ''; 
    phone: string= ''; 
    email: string= ''; 
    address: string= ''; 
    latitude: string= ''; 
    longitude: string= ''; 
}